using_mailslurp_with_capybara/python/linux-requirements.txt
boto3==1.18.1
botocore==1.21.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
idna==3.2
jmespath==0.10.0
lorem==0.1.1
Pillow==8.3.1
protobuf==3.17.3
PyJWT==2.1.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
tornado==6.1
urllib3==1.26.6
watchdog==2.1.3
